+++ examples/example-twitter-server/src/test/scala/com/twitter/finatra/sample/ExampleTwitterServerStartupTest.scala
@@ -1,0 +1,16 @@
+package com.twitter.finatra.sample
+
+import com.google.inject.Stage
+import com.twitter.inject.server.{EmbeddedTwitterServer, FeatureTest}
+
+class ExampleTwitterServerStartupTest extends FeatureTest {
+
+  val server = new EmbeddedTwitterServer(
+    twitterServer = new ExampleTwitterServer,
+    stage = Stage.PRODUCTION
+  )
+
+  test("Server#startup") {
+    server.assertHealthy()
+  }
+}