+++ http/src/main/scala/com/twitter/finatra/http/filters/TraceIdMDCFilter.scala
@@ -2,6 +2,7 @@ package com.twitter.finatra.http.filters
 
 import com.twitter.finagle.tracing.Trace
 import com.twitter.finagle.{Service, SimpleFilter}
+import com.twitter.util.Future
 import javax.inject.Singleton
 import org.slf4j.MDC
 
@@ -12,7 +13,7 @@ import org.slf4j.MDC
 @Singleton
 class TraceIdMDCFilter[Req, Rep] extends SimpleFilter[Req, Rep] {
 
-  override def apply(request: Req, service: Service[Req, Rep]) = {
+  override def apply(request: Req, service: Service[Req, Rep]): Future[Rep] = {
     MDC.put("traceId", Trace.id.traceId.toString())
     service(request)
   }