+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/mdc/main/AnotherService.scala
@@ -1,0 +1,9 @@
+package com.twitter.finatra.http.tests.integration.mdc.main
+
+import com.twitter.util.Future
+
+class AnotherService {
+
+  def sendNotification(userPreferences: UserPreferences): Future[NotificationEvent] =
+    Future.value(NotificationEvent(userPreferences))
+}