+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/mdc/main/EventService.scala
@@ -1,0 +1,8 @@
+package com.twitter.finatra.http.tests.integration.mdc.main
+
+import com.twitter.util.Future
+
+class EventService {
+
+  def registerEvent(notificationEvent: NotificationEvent): Future[Unit] = Future.Done
+}