+++ http/src/test/scala/com/twitter/finatra/http/tests/integration/mdc/main/NotificationEvent.scala
@@ -1,0 +1,5 @@
+package com.twitter.finatra.http.tests.integration.mdc.main
+
+case class NotificationEvent(userPreferences: UserPreferences) {
+  def sent: Boolean = true
+}